// src/app/extends/extend-company-labels.enum.ts
export enum ExtendCompanyLabels {
  CompanyNameLabel = 'أسم الشركة',
  logoNameLabel = 'لوجو الشركة',
  urlMapNameLabel = 'موقع الشركة',
  briefNameLabel = 'نبذه عن الشركة',
  categoryNameLabel = 'الفئة',
  specializationNameLabel = 'التخصص',
  idCompanyNameLabel = 'رقم رخصة الشركة',
  companyImageNameLabel = 'صور الشركة',
  websiteUrlNameLabel = 'رابط الموقع',
  facebookUrlNameLabel = 'رابط الفيس بوك',
  twitterUrlNameLabel = 'رابط تويتر',
  instagramUrlNameLabel = 'رابط انستجرام',
  companyRateNameLabel = 'تقييم الشركة',
  companyViewNameLabel = 'عدد المشاهدات',
  companyCity = 'المدينة',
  companyphone = 'رقم الهاتف',
  companyEmail = 'البريد الإلكتروني',
  companyAddress = 'عنوان الشركة',
  companyLinks = 'روابط مشاركة'
}