// src/app/extends/extend-add-project-labels.enum.ts
export enum ExtendAddProjectLabels {
  projectNameLabel = 'أسم المشروع',
  briefProjectNameLabel = 'نبذه عن المشروع',
  projectAddress = 'عنوان المشروع',
  projectImagesNameLabel = 'صور المشروع',
  naturalProjectImagesNameLabel = 'صور من ارض الواقع'
}